+++ docs/Changelog.md
@@ -2,6 +2,17 @@
 
 ## Next version
 
+- `ThemeCallback.getColor` method [#1251](https://github.com/mapsforge/vtm/pull/1251)
+- Minor improvements and bug fixes
+- [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.27.0)
+
+## Version 0.26.1 (2025-09-15)
+
+- Minor improvements and bug fixes
+- [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.26.1)
+
+## Version 0.26.0 (2025-09-14)
+
 - Repeat proximity for labels with same text [#1232](https://github.com/mapsforge/vtm/pull/1232)
   - `Parameters.REPEAT_PROXIMITY_SQR`
 - `ThemeCallback.getText` method [#1225](https://github.com/mapsforge/vtm/pull/1225)